Fixed payment amount bug, code refactor
